Revert "AuthN: move oauth token hook into session client (#76688)"
This reverts commit 455cede6992cfad57f1c5b1749c9454d1475dba6.
