+++ pkg/services/authn/authnimpl/sync/user_sync.go
@@ -394,5 +394,4 @@ func syncSignedInUserToIdentity(usr *user.SignedInUser, identity *authn.Identity
 	identity.LastSeenAt = usr.LastSeenAt
 	identity.IsDisabled = usr.IsDisabled
 	identity.IsGrafanaAdmin = &usr.IsGrafanaAdmin
-	identity.AuthenticatedBy = usr.AuthenticatedBy
 }
